chore(jit-html): cleanup/reorganize tests
