+++ packages/jit-html/test/integration/kitchen-sink.spec.ts
@@ -1,0 +1,360 @@
+import { IAttributeParser, ResourceModel } from '@aurelia/jit';
+import { RuntimeCompilationResources } from '@aurelia/kernel';
+import {
+  Aurelia,
+  CustomElementResource,
+  IExpressionParser,
+  ILifecycle,
+  INodeSequence,
+  ISignaler,
+  LifecycleFlags } from '@aurelia/runtime';
+import { expect } from 'chai';
+import { NodeSequenceFactory } from '../../../runtime-html/src/dom';
+import { TemplateBinder } from '../../src/index';
+import { TestContext } from '../util';
+
+const spec = 'kitchen-sink';
+
+// TemplateCompiler - integration with various different parts
+describe(spec, () => {
+  it('startup with App type', () => {
+    const ctx = TestContext.createHTMLTestContext();
+    const component = CustomElementResource.define({ name: 'app', template: `<template>\${message}</template>` }, class { public message = 'Hello!'; });
+    const host = ctx.createElement('div');
+    const au = new Aurelia(ctx.container).register().app({ host, component }).start();
+    expect(host.textContent).to.equal('Hello!');
+    au.stop();
+    expect(host.textContent).to.equal('');
+    au.start();
+    expect(host.textContent).to.equal('Hello!');
+    au.stop();
+    expect(host.textContent).to.equal('');
+  });
+
+
+  it('signaler', async () => {
+
+    const items = [0, 1, 2];
+    const App = CustomElementResource.define({
+      name: 'app',
+      template: `<template><div repeat.for="i of 3">\${items[i] & signal:'updateItem'}</div></template>`
+    },                                       class {
+      public items = items;
+    });
+
+    const ctx = TestContext.createHTMLTestContext();
+    const signaler = ctx.container.get(ISignaler);
+    const lifecycle = ctx.lifecycle;
+    const au = new Aurelia(ctx.container);
+
+    const host = ctx.createElement('div');
+    const component = new App();
+
+    au.app({ host, component });
+
+    au.start();
+
+    expect(host.textContent).to.equal('012');
+
+    items[0] = 2;
+
+    lifecycle.processFlushQueue(LifecycleFlags.none);
+
+    expect(host.textContent).to.equal('012');
+
+    signaler.dispatchSignal('updateItem', LifecycleFlags.fromFlush);
+
+    expect(host.textContent).to.equal('212');
+
+  });
+
+  it('signaler + oneTime', async () => {
+
+    const items = [0, 1, 2];
+    const App = CustomElementResource.define({
+      name: 'app',
+      template: `<template><div repeat.for="i of 3">\${items[i] & signal:'updateItem' & oneTime}</div></template>`
+    },                                       class {
+      public items = items;
+    });
+
+    const ctx = TestContext.createHTMLTestContext();
+    const signaler = ctx.container.get(ISignaler);
+    const lifecycle = ctx.lifecycle;
+    const au = new Aurelia(ctx.container);
+
+    const host = ctx.createElement('div');
+    const component = new App();
+
+    au.app({ host, component });
+
+    au.start();
+
+    expect(host.textContent).to.equal('012');
+
+    items[0] = 2;
+
+    lifecycle.processFlushQueue(LifecycleFlags.none);
+
+    expect(host.textContent).to.equal('012');
+
+    signaler.dispatchSignal('updateItem', LifecycleFlags.fromFlush);
+
+    expect(host.textContent).to.equal('212');
+
+  });
+
+  it('render hook', async () => {
+
+    const ctx = TestContext.createHTMLTestContext();
+    const App = CustomElementResource.define({
+      name: 'app',
+      template: `<template></template>`
+    },                                       class {
+      public $nodes: INodeSequence;
+      public render() {
+        this.$nodes = new NodeSequenceFactory(ctx.dom, 'foo').createNodeSequence();
+      }
+    });
+
+    const au = new Aurelia(ctx.container);
+
+    const host = ctx.createElement('div');
+    const component = new App();
+
+    au.app({ host, component });
+
+    au.start();
+
+    expect(host.textContent).to.equal('foo');
+
+  });
+});
+
+describe('xml node compiler tests', () => {
+  // TODO: add some content assertions and verify different kinds of xml compilation
+  // (for now these tests are just to ensure the binder doesn't hang or crash when given "unusual" node types)
+  const markups = [
+    '<?xml?>',
+    '<?xml version="1.0" encoding="utf-8"?>',
+    '<?xml?>\n<a/>',
+    '<?xml?>\n<a>\n\v<b/>\n</a>',
+    '<?go there?>',
+    '<?go there?><?come here?>',
+    '<!-- \t Hello, World! \t -->',
+    '<!-- \t Hello \t -->\n<!-- \t World \t -->',
+    '<![CDATA[ \t <foo></bar> \t ]]>',
+    '<![CDATA[ \t data]]><![CDATA[< > " and & \t ]]>',
+    '<!DOCTYPE note [\n<!ENTITY foo "baa">]>',
+    '<a/>',
+    '<a/>\n<a/>',
+    '<a/>\n<b/>',
+    '<a x="hello"/>',
+    '<a x="1.234" y="It\'s"/>',
+    '<a> \t Hi \t </a>',
+    '<a>  Hi \n There \t </a>',
+    '<a>\n\v<b/>\n</a>',
+    '<a>\n\v<b>\n\v\v<c/>\n\v</b>\n</a>'
+  ];
+
+  for (const markup of markups) {
+    const escaped = markup.replace(/\b/g, '\\b').replace(/\t/g, '\\t').replace(/\n/g, '\\n').replace(/\v/g, '\\v').replace(/\f/g, '\\f').replace(/\r/g, '\\r');
+    it(escaped, () => {
+      const ctx = TestContext.createHTMLTestContext();
+      const parser = new ctx.DOMParser();
+      const doc = parser.parseFromString(markup, 'application/xml');
+      const fakeSurrogate = { firstChild: doc, attributes: [] };
+
+      const binder = new TemplateBinder(
+        ctx.dom,
+        new ResourceModel(new RuntimeCompilationResources(ctx.container)),
+        ctx.container.get(IAttributeParser),
+        ctx.container.get(IExpressionParser)
+      );
+
+      const result = binder.bind(fakeSurrogate as any);
+      expect(result.physicalNode).to.equal(fakeSurrogate);
+    });
+  }
+});
+
+describe('dependency injection', () => {
+
+  it('register local dependencies ', () => {
+    const Foo = CustomElementResource.define(
+      {
+        name: 'foo',
+        template: 'bar'
+      },
+      null
+    );
+    const App = CustomElementResource.define(
+      {
+        name: 'app',
+        template: '<foo></foo>',
+        dependencies: [Foo]
+      },
+      null
+    );
+
+    const ctx = TestContext.createHTMLTestContext();
+    ctx.container.register(Foo);
+    const au = new Aurelia(ctx.container);
+
+    const host = ctx.createElement('div');
+    const component = new App();
+
+    au.app({ host, component });
+
+    au.start();
+
+    expect(host.textContent).to.equal('bar');
+  });
+});
+
+
+// commented out code left here intentionally, serves as a staring point for template controller tests
+
+// describe.only('test', () => {
+//   it.only('if', () => {
+//     enableTracing();
+//     Tracer.enableLiveLogging(SymbolTraceWriter);
+//     const container = BasicConfiguration.createContainer();
+//     const dom = new HTMLDOM(document);
+//     Registration.instance(IDOM, dom).register(container, IDOM);
+//     const host = document.createElement('div');
+//     const hosthost = document.createElement('div');
+//     hosthost.appendChild(host);
+//     const App = CustomElementResource.define(
+//       {
+//         name: 'app',
+//         template: '<div if.bind="value">${ifText}</div><div else>${elseText}</div>'
+//       },
+//       class {
+//         ifText = 'foo';
+//         elseText = 'bar';
+//       }
+//     );
+//     const component = new App();
+
+//     const lifecycle = container.get(ILifecycle);
+//     const re = container.get(IRenderingEngine);
+//     const pl = container.get(IProjectorLocator);
+
+//     component.$hydrate(dom, pl, re, host, null);
+
+//     component['value'] = false;
+//     component.$bind(LifecycleFlags.none);
+
+//     component['value'] = true;
+
+//     component.$bind(LifecycleFlags.none);
+
+//     component.$attach(LifecycleFlags.none);
+
+//     disableTracing();
+//     expect(host.textContent).to.equal('bar')
+//   });
+
+//   it.only('if2', () => {
+//     enableTracing();
+//     Tracer.enableLiveLogging(SymbolTraceWriter);
+//     // common stuff
+//     const container = AuDOMConfiguration.createContainer();
+//     const dom = container.get<AuDOM>(IDOM);
+//     const observerLocator = container.get(IObserverLocator);
+//     const lifecycle = container.get(ILifecycle);
+
+//     const location = AuNode.createRenderLocation();
+//     const host = AuNode.createHost().appendChild(location.$start).appendChild(location);
+
+//     const ifPropName = 'ifValue';
+//     const elsePropName = 'elseValue';
+//     const ifText = 'foo';
+//     const elseText = 'bar';
+
+//     const ifTemplate: ITemplate<AuNode> = {
+//       renderContext: null as any,
+//       dom: null as any,
+//       render(renderable) {
+//         const text = AuNode.createText();
+//         const wrapper = AuNode.createTemplate().appendChild(text);
+
+//         const nodes = new AuNodeSequence(dom, wrapper);
+//         const binding = new Binding(new AccessScope(ifPropName), text, 'textContent', BindingMode.toView, observerLocator, container);
+
+//         (renderable as Writable<typeof renderable>).$nodes = nodes;
+//         addBinding(renderable, binding);
+//       }
+//     };
+
+//     const elseTemplate: ITemplate<AuNode> = {
+//       renderContext: null as any,
+//       dom: null as any,
+//       render(renderable) {
+//         const text = AuNode.createText();
+//         const wrapper = AuNode.createTemplate().appendChild(text);
+
+//         const nodes = new AuNodeSequence(dom, wrapper);
+//         const binding = new Binding(new AccessScope(elsePropName), text, 'textContent', BindingMode.toView, observerLocator, container);
+
+//         (renderable as Writable<typeof renderable>).$nodes = nodes;
+//         addBinding(renderable, binding);
+//       }
+//     };
+
+//     //@ts-ignore
+//     const ifFactory = new ViewFactory<AuNode>('if-view', ifTemplate, lifecycle);
+//     //@ts-ignore
+//     const elseFactory = new ViewFactory<AuNode>('else-view', elseTemplate, lifecycle);
+
+//     const sut = new If<AuNode>(ifFactory, location, new CompositionCoordinator(lifecycle));
+//     const elseSut = new Else<AuNode>(elseFactory);
+//     elseSut.link(sut);
+
+//     (sut as Writable<If>).$scope = null;
+//     (elseSut as Writable<Else>).$scope = null;
+
+//     //@ts-ignore
+//     const ifBehavior = RuntimeBehavior.create(If);
+//     //@ts-ignore
+//     ifBehavior.applyTo(sut, lifecycle);
+
+//     //@ts-ignore
+//     const elseBehavior = RuntimeBehavior.create(Else);
+//     //@ts-ignore
+//     elseBehavior.applyTo(elseSut, lifecycle);
+
+//     let firstBindInitialNodesText: string;
+//     let firstBindFinalNodesText: string;
+//     let secondBindInitialHostsText: string;
+//     let secondBindFinalNodesText: string;
+//     let firstAttachInitialHostText: string = 'foo';
+//     let firstAttachFinalHostText: string;
+//     let secondAttachInitialHostText: string;
+//     let secondAttachFinalHostText: string;
+
+//     // -- Round 1 --
+
+//     const ctx = BindingContext.create({
+//       [ifPropName]: ifText,
+//       [elsePropName]: elseText
+//     });
+//     let scope = Scope.create(ctx);
+
+//     sut.value = false;
+
+//     sut.$bind(LifecycleFlags.none, scope);
+//     if (false) {
+//       scope = Scope.create(ctx);
+//     }
+//     sut.value = true;
+//     sut.$bind(LifecycleFlags.none, scope);
+
+//     sut.$attach(LifecycleFlags.none);
+
+//     expect(host.textContent).to.equal(firstAttachInitialHostText, 'host.textContent #1');
+//     disableTracing();
+
+//   });
+// });
